Fix circular dependency that was breaking tests
