+++ packages/provider/src/index.ts
@@ -1,4 +1,5 @@
-export { SDKBase, createSDK, MagicSDKAdditionalConfiguration } from './core/sdk';
+export { SDKBase, MagicSDKAdditionalConfiguration } from './core/sdk';
+export { createSDK } from './core/sdk-environment';
 export { PayloadTransport } from './core/payload-transport';
 export { ViewController } from './core/view-controller';
 export * from './core/sdk-exceptions';
