+++ packages/provider/test/spec/modules/base-module/constructor.spec.ts
@@ -3,18 +3,7 @@
 import browserEnv from '@ikscodes/browser-env';
 import test from 'ava';
 import { createMagicSDK } from '../../../factories';
-
-/**
- * We have a circular dependency breaking test code when the `BaseModule`
- * constructor is referenced. Rather than refactor the SDK code, it was quicker
- * to fix the issue with JS getters.
- */
-const ModuleCtors = {
-  get BaseModule() {
-    return (require('../../../../src/modules/base-module') as typeof import('../../../../src/modules/base-module'))
-      .BaseModule;
-  },
-};
+import { BaseModule } from '../../../../src/modules/base-module';
 
 test.beforeEach(t => {
   browserEnv.restore();
@@ -23,7 +12,7 @@ test.beforeEach(t => {
 test.serial('Initialize `BaseModule`', t => {
   const sdk = createMagicSDK();
 
-  const baseModule = new ModuleCtors.BaseModule(sdk);
+  const baseModule = new BaseModule(sdk);
 
   const transportA = (baseModule as any).transport;
   const transportB = (baseModule as any).sdk.transport;
@@ -31,7 +20,7 @@ test.serial('Initialize `BaseModule`', t => {
   const overlayA = (baseModule as any).overlay;
   const overlayB = (baseModule as any).sdk.overlay;
 
-  t.true(baseModule instanceof ModuleCtors.BaseModule);
+  t.true(baseModule instanceof BaseModule);
   t.is(transportA, transportB);
   t.is(overlayA, overlayB);
 });
